requirements.txt
python-telegram-bot[job-queue]==20.7
gspread==6.0.2
google-auth==2.26.2